components/ui 이동 및 shared/components 삭제
