+++ package.json
@@ -49,6 +49,7 @@
     "react-toastify": "^11.0.5",
     "shadcn": "^4.10.0",
     "socket.io-client": "^4.8.1",
+    "tailwind-animations": "^1.0.1",
     "tailwind-merge": "^3.6.0",
     "tailwindcss": "^4.1.18",
     "tw-animate-css": "^1.4.0",
