+++ src/features/lobby/components/PlayerInfo.tsx
@@ -1,5 +1,5 @@
 import { motion } from "motion/react";
-import { Avatar } from "@/shared/components/Avatar";
+import { Avatar } from "@/components/ui/Avatar";
 import { animalList } from "@/shared/constants/avatarCandidates";
 import Subtitle from "./Subtitle";
 import { ImageManager } from "@/shared/services/ImageManger";
