+++ src/index.css
@@ -1,8 +1,16 @@
-@import url('https://fonts.googleapis.com/css2?family=Luckiest+Guy&display=swap');
+@import url("https://fonts.googleapis.com/css2?family=Luckiest+Guy&display=swap");
 @import "tailwindcss";
 @import "tw-animate-css";
 @import "shadcn/tailwind.css";
 @custom-variant dark (&:is(.dark *));
+
+@font-face {
+  font-family: "CookieRun";
+  src: url("https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts_2001@1.1/CookieRun-Regular.woff")
+    format("woff");
+  font-weight: normal;
+  font-display: swap;
+}
 
 html {
   font-size: calc(10px + 0.5vw);
@@ -30,6 +38,7 @@ body {
   height: 100%;
   overflow: hidden;
   background-color: var(--background);
+  background-image: url("../assets/backgrounds/background.svg");
   /* background:
     radial-gradient(
       circle at top,
@@ -44,19 +53,19 @@ body {
   animation: move-pattern 10s linear infinite; */
 }
 @layer base {
-  h1 { 
-    @apply text-primary text-[400%] rotate-5 text-retro-stroke;
+  h1 {
+    @apply text-primary text-[500%] rotate-5 text-retro-stroke;
   }
 }
 
-/* 까만 테두리 그림자 조합 */ 
+/* 까만 테두리 그림자 조합 */
 @utility text-retro-stroke {
   /* 1. 글자 겉면 테두리 두께와 색상 */
   -webkit-text-stroke: 2px var(--foreground);
   /* 2. 테두리가 글자 안쪽(Fill)을 파먹지 않도록 바깥으로만 밀어냄 (초특급 꿀팁!) */
-  paint-order: stroke fill; 
+  paint-order: stroke fill;
   /* 3. 테두리 바깥으로 떨어지는 레트로 그림자 */
-  text-shadow: 6px 6px 0px var(--foreground); 
+  text-shadow: 6px 6px 0px var(--foreground);
 }
 /* 브루탈 박스 */
 @utility brutal-box {
@@ -105,7 +114,7 @@ body {
 
 @theme inline {
   --font-heading: var(--font-sans);
-  --font-sans: "Luckiest Guy", sans-serif;
+  --font-sans: "Luckiest Guy", "CookieRun";
   --color-sidebar-ring: var(--sidebar-ring);
   --color-sidebar-border: var(--sidebar-border);
   --color-sidebar-accent-foreground: var(--sidebar-accent-foreground);
@@ -173,7 +182,7 @@ body {
   --radius: 0;
   --sidebar: oklch(0.985 0.001 106.423);
   --sidebar-foreground: oklch(0.147 0.004 49.25);
-  --sidebar-primary: oklch(0.546 0.245 262.881);
+  --sidebar-primary: oklch(87.769% 0.16641 95.738);
   --sidebar-primary-foreground: oklch(0.97 0.014 254.604);
   --sidebar-accent: oklch(0.97 0.001 106.424);
   --sidebar-accent-foreground: oklch(0.216 0.006 56.043);
@@ -188,7 +197,7 @@ body {
   --card-foreground: oklch(0.985 0.001 106.423);
   --popover: oklch(0.216 0.006 56.043);
   --popover-foreground: oklch(0.985 0.001 106.423);
-  --primary: oklch(0.424 0.199 265.638);
+  --primary: oklch(87.769% 0.16641 95.738);
   --primary-foreground: oklch(0.97 0.014 254.604);
   --secondary: oklch(0.274 0.006 286.033);
   --secondary-foreground: oklch(0.985 0 0);
@@ -207,7 +216,7 @@ body {
   --chart-5: oklch(0.47 0.157 37.304);
   --sidebar: oklch(0.216 0.006 56.043);
   --sidebar-foreground: oklch(0.985 0.001 106.423);
-  --sidebar-primary: oklch(0.623 0.214 259.815);
+  --sidebar-primary: oklch(87.769% 0.16641 95.738);
   --sidebar-primary-foreground: oklch(0.97 0.014 254.604);
   --sidebar-accent: oklch(0.268 0.007 34.298);
   --sidebar-accent-foreground: oklch(0.985 0.001 106.423);
