+++ src/layouts/TicTacToe.tsx
@@ -1,5 +1,5 @@
 import Players from "../features/game/components/Players";
-import Countdown from "@/shared/components/Countdown";
+import Countdown from "@/components/ui/Countdown";
 import { useGameStore } from "@/stores/useGameStore";
 import { useBackExitModal } from "@/shared/hooks/useBackExitModal";
